+++ Nudge/Preferences/PreferencesStructure.swift
@@ -66,7 +66,7 @@ extension NudgePreferences {
 // MARK: - OptionalFeatures
 struct OptionalFeatures: Codable {
     var acceptableApplicationBundleIDs, acceptableAssertionApplicationNames: [String]?
-    var acceptableAssertionUsage, acceptableCameraUsage, acceptableUpdatePreparingUsage, acceptableScreenSharingUsage, aggressiveUserExperience, aggressiveUserFullScreenExperience, asynchronousSoftwareUpdate, attemptToBlockApplicationLaunches, attemptToCheckForSupportedDevice, attemptToFetchMajorUpgrade: Bool?
+    var acceptableAssertionUsage, acceptableCameraUsage, acceptableUpdatePreparingUsage, acceptableScreenSharingUsage, aggressiveUserExperience, aggressiveUserFullScreenExperience, allowMovableWindowDuringAggressiveUserExperience, asynchronousSoftwareUpdate, attemptToBlockApplicationLaunches, attemptToCheckForSupportedDevice, attemptToFetchMajorUpgrade: Bool?
     var blockedApplicationBundleIDs: [String]?
     var customSOFAFeedURL: String?
     var disableNudgeForStandardInstalls, disableSoftwareUpdateWorkflow, enforceMinorUpdates, honorFocusModes, honorCycleTimersOnExit: Bool?
@@ -101,6 +101,7 @@ extension OptionalFeatures {
         acceptableScreenSharingUsage: Bool? = nil,
         aggressiveUserExperience: Bool? = nil,
         aggressiveUserFullScreenExperience: Bool? = nil,
+        allowMovableWindowDuringAggressiveUserExperience: Bool? = nil,
         asynchronousSoftwareUpdate: Bool? = nil,
         attemptToBlockApplicationLaunches: Bool? = nil,
         attemptToCheckForSupportedDevice: Bool? = nil,
@@ -125,6 +126,7 @@ extension OptionalFeatures {
             acceptableScreenSharingUsage: acceptableScreenSharingUsage ?? self.acceptableScreenSharingUsage,
             aggressiveUserExperience: aggressiveUserExperience ?? self.aggressiveUserExperience,
             aggressiveUserFullScreenExperience: aggressiveUserFullScreenExperience ?? self.aggressiveUserFullScreenExperience,
+            allowMovableWindowDuringAggressiveUserExperience: allowMovableWindowDuringAggressiveUserExperience ?? self.allowMovableWindowDuringAggressiveUserExperience,
             asynchronousSoftwareUpdate: asynchronousSoftwareUpdate ?? self.asynchronousSoftwareUpdate,
             attemptToBlockApplicationLaunches: attemptToBlockApplicationLaunches ?? self.attemptToBlockApplicationLaunches,
             attemptToCheckForSupportedDevice: attemptToCheckForSupportedDevice ?? self.attemptToCheckForSupportedDevice,