+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Requires macOS 12.0 and higher.
 ### Added
 - `showRequiredOSVersion` is a new key, defaulted to `False`. This currently has no impact to standardMode
 - simpleMode now honors `showRequiredOSVersion` and `showRequiredDate` allowing you to give slightly more information to the user while keeping a minimal UI
+- `allow​Movable​Window​During​Aggressive​User​Experience` key is added
+  - Addresses [671](https://github.com/macadmins/nudge/issues/671)
 
 ### Changed
  - Swedish translation for Required Date is more culturally appropriate